Remove mapChunks, rename initializer parameters
