+++ Sources/OpenAPIRuntime/Conversion/CurrencyExtensions.swift
@@ -65,35 +65,6 @@ extension HTTPRequest {
     }
 }
 
-extension HTTPRequest {
-
-    @_spi(Generated)
-    public init(path: String, method: Method) {
-        self.init(method: method, scheme: nil, authority: nil, path: path)
-    }
-
-    @_spi(Generated)
-    public var query: Substring? {
-        guard let path else {
-            return nil
-        }
-        guard let queryStart = path.firstIndex(of: "?") else {
-            return nil
-        }
-        let queryEnd = path.firstIndex(of: "#") ?? path.endIndex
-        let query = path[path.index(after: queryStart)..<queryEnd]
-        return query
-    }
-}
-
-extension HTTPResponse {
-
-    @_spi(Generated)
-    public init(statusCode: Int) {
-        self.init(status: .init(code: statusCode))
-    }
-}
-
 extension Converter {
 
     // MARK: Converter helpers
